requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.7.0
click==8.1.8
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
pydantic==2.10.4
pydantic_core==2.27.2
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0

httpx~=0.28.1
python-dotenv~=1.0.1
python-jose~=3.3.0
bcrypt~=4.2.1
Jinja2~=3.1.5